fix(tresor): return error on failure to decode cert or key
Returns an error when decoding the ca cert chain or the ca private
key. Currently, if the ca cert or private key contained in the ca
bundle secret are configured incorrectly and Tresor is the cert
provider, a segmentation fault will occur when trying to issue
a cert due to a nil pointer dereference. After a err occurs when
decoding the ca cert or private key the function should return the
error to avoid the noisy seg fault.

Signed-off-by: jaellio <jaellio@microsoft.com>